Share part of the global_asm!() implementation between cg_ssa and cg_clif
